// ebegu-web/src/models/enums/TSWizardSubStepName.ts
/*
 * Copyright (C) 2020 DV Bern AG, Switzerland
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

export enum TSWizardSubStepName {
    KEIN_WEITERER_SUBSTEP = 'KEIN_WEITERER_SUBSTEP',
    FINANZIELLE_SITUATION_START = 'FINANZIELLE_SITUATION_START',
    FINANZIELLE_SITUATON_GS1 = 'FINANZIELLE_SITUATON_GS1',
    FINANZIELLE_SITUATON_GS2 = 'FINANZIELLE_SITUATON_GS2',
    FINANZIELLE_SITUATION_RESULTATE = 'FINANZIELLE_SITUATION_RESULTATE',
    FINANZIELLE_SITUATION_SOZIALHILFE = 'FINANZIELLE_SITUATION_SOZIALHILFE',
    FINANZIELLE_SITUATION_SOZIALHILFE_DETAIL = 'FINANZIELLE_SITUATION_SOZIALHILFE_DETAIL',
}